+++ samples/MortgageApplication/cobol/epscmort.cbl
@@ -1,7 +1,7 @@
        ID DIVISION.
        PROGRAM-ID. EPSCMORT.
       *    THIS DEMONSTRATES CICS/DEBUG           - EPSDEMOS 2008
-      * changed July 13, 2020 21:25
+      * changed July 14, 2020 09:07
       *    THIS PROGRAM WILL RECEIVE A DATE AND COVERT THE DATE TO
       *    AN INTEGER IN A CALLED PROGRAM TO DETERMINE DAYS FROM
       *    CURRENT DATE.
